+++ src/app/pages/archive/archive.module.ts
@@ -1,18 +1,18 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {ReusableComponentsModule} from "../../../shared/reusablecomponents/reusable-components.module";
+import {ReusableComponentsModule} from "../../shared/reusablecomponents/reusable-components.module";
 import {ArchiveRouting} from "./archive.routing";
 import {ArchiveComponent} from "./archive.component";
 import {NCTEPoliciesComponent} from "./policies/ncte-policies.component";
 import {NCTEFundingComponent} from "./funding/ncte-funding.component";
 import {HighchartsChartModule} from "highcharts-angular";
-import {DataService} from "../../../services/data.service";
+import {DataService} from "../../services/data.service";
 import {CountriesTableComponent} from "./countries-table.component";
-import {DataHandlerService} from "../../../services/data-handler.service";
+import {DataHandlerService} from "../../services/data-handler.service";
 import {NCTEMandate} from "./mandate/ncte-mandate";
 import {NCTEOpenAccessComponent} from "./open-access/ncte-open-access.component";
 import {NCTEMonitoringComponent} from "./monitoring/ncte-monitoring.component";
-import {StakeholdersService} from "../../../services/stakeholders.service";
+import {StakeholdersService} from "../../services/stakeholders.service";
 
 @NgModule({
   declarations: [
@@ -37,7 +37,9 @@ import {StakeholdersService} from "../../../services/stakeholders.service";
   ],
   exports: [
     NCTEPoliciesComponent,
-    CountriesTableComponent
+    CountriesTableComponent,
+    NCTEFundingComponent,
+    NCTEMandate
   ]
 })
