+++ src/app/pages/archive/mandate/ncte-mandate.ts
@@ -1,10 +1,10 @@
 import {Component} from "@angular/core";
-import {environment} from "../../../../../environments/environment";
-import {CountryTableData} from "../../../../domain/country-table-data";
+import {environment} from "../../../../environments/environment";
+import {CountryTableData} from "../../../domain/country-table-data";
 import {DomSanitizer, SafeResourceUrl} from "@angular/platform-browser";
-import {DataService} from "../../../../services/data.service";
-import {DataHandlerService} from "../../../../services/data-handler.service";
-import {CategorizedAreaData} from "../../../../domain/categorizedAreaData";
+import {DataService} from "../../../services/data.service";
+import {DataHandlerService} from "../../../services/data-handler.service";
+import {CategorizedAreaData} from "../../../domain/categorizedAreaData";
 
 
 @Component({
