+++ src/app/pages/archive/monitoring/ncte-monitoring.component.ts
@@ -1,9 +1,9 @@
 import {Component, OnInit} from "@angular/core";
-import {DataService} from "../../../../services/data.service";
-import {CountryTableData} from "../../../../domain/country-table-data";
-import {DataHandlerService} from "../../../../services/data-handler.service";
+import {DataService} from "../../../services/data.service";
+import {CountryTableData} from "../../../domain/country-table-data";
+import {DataHandlerService} from "../../../services/data-handler.service";
 import {DomSanitizer, SafeResourceUrl} from "@angular/platform-browser";
-import {environment} from "../../../../../environments/environment";
+import {environment} from "../../../../environments/environment";
 
 @Component({
   selector: 'app-ncte-monitoring',
