+++ src/app/pages/national-contributions-to-eosc-dashboard/national-contributions-to-eosc-dashboard.component.html
@@ -11,7 +11,6 @@
             <ul class="uk-navbar-nav">
               <li [routerLinkActive]="['uk-active']"><a [routerLink]="'./policies'" class="sub-menu-item">Policies</a></li>
               <li [routerLinkActive]="['uk-active']"><a [routerLink]="'./practices'" class="sub-menu-item">Practices</a></li>
-              <li [routerLinkActive]="['uk-active']"><a [routerLink]="'./archive'" class="sub-menu-item">Archive</a></li>
             </ul>
           </div>
         </nav>
