+++ src/app/pages/national-contributions-to-eosc-dashboard/national-contributions-to-eosc-dashboard.module.ts
@@ -9,7 +9,7 @@ import {DataHandlerService} from "../../services/data-handler.service";
 import {StakeholdersService} from "../../services/stakeholders.service";
 import {PoliciesComponent} from "./policies/policies.component";
 import {PracticesComponent} from "./practices/practices.component";
-import {ArchiveModule} from "./archive/archive.module";
+import {ArchiveModule} from "../archive/archive.module";
 
 @NgModule({
   declarations: [
