+++ src/app/pages/national-contributions-to-eosc-dashboard/national-contributions-to-eosc-dashboard.routing.ts
@@ -23,10 +23,6 @@ const nationalContributionsToEOSCDashboardRoutes: Routes = [
       {
         path: 'practices',
         component: PracticesComponent,
-      },
-      {
-        path: 'archive',
-        loadChildren: () => import('./archive/archive.module').then(m => m.ArchiveModule)
       }
     ],
     runGuardsAndResolvers: 'always'
