+++ src/app/pages/national-contributions-to-eosc-dashboard/policies/policies.component.html
@@ -22,89 +22,8 @@
       </div>
       <div class="uk-width-1-1">
         <ul id="tabs" class="uk-switcher">
-          <li class="uk-margin-medium-top">
-            <ul class="uk-flex-inline uk-flex-left uk-subnav uk-subnav-pill-alt uk-margin-remove"
-                data-uk-switcher="{connect:'#map-vs-table-1', animation: 'fade'}">
-              <li>
-                <a class="uk-margin-remove-bottom"><span>Graph view</span></a>
-              </li>
-              <li>
-                <a class="uk-margin-remove-bottom"><span>Table view</span></a>
-              </li>
-            </ul>
-            <ul id="map-vs-table-1" class="uk-switcher uk-margin">
-              <li>
-                <div class="uk-card uk-card-default uk-card-body">
-                  <div class="uk-grid uk-grid-match">
-                    <div class="uk-width-1-3 uk-margin-medium-top">
-                      <div class="uk-text-meta uk-margin-bottom uk-text-italic">
-                        Select one of the following to change the view of the map
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" checked (click)="createMapDataset(0,1)">
-                        <label>There are one or more policies relevant for the EOSC in place</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(1, 1)">
-                        <label>Policy in planning</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(2, 1)">
-                        <label>One or more of the open science policies explicitly mentions EOSC</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(3, 1)">
-                        <label>Policy addresses Open access to data, data management and/or FAIR</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(4, 1)">
-                        <label>Policy addresses FAIRisation of data</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(5, 1)">
-                        <label>Policy addresses Open access to software</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(6, 1)">
-                        <label>Policy addresses Preservation and reuse of scientific information</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(7, 1)">
-                        <label>Policy addresses Infrastructures that include aspects of open science</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(8, 1)">
-                        <label>Policy addresses Skills and competencies</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(9, 1)">
-                        <label>Policy addresses Incentives and rewards</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(10, 1)">
-                        <label>Policy addresses Citizen science</label>
-                      </div>
-                      <div class="uk-flex uk-flex-top uk-margin-small-bottom">
-                        <input type="radio" name="map1" (click)="createMapDataset(11, 1)">
-                        <label>Other</label>
-                      </div>
-                    </div>
-                    <div class="uk-width-2-3" style="height: 700px">
-                      <app-highcharts-category-map [mapData]="questionsDataArray[1]" [subtitle]="mapSubtitles[0]"
-                                                   [title]="'EOSC-relevant policies in place at national or regional level'">
-                      </app-highcharts-category-map>
-                    </div>
-                  </div>
-                </div>
-              </li>
-              <li>
-                <div class="uk-card uk-card-default uk-card-body">
-                  <app-countries-table *ngIf="tableAbsoluteDataArray[1]" [countries]="tableAbsoluteDataArray[1]"
-                                       [tableType]="'policies'">
-                  </app-countries-table>
-                </div>
-              </li>
-            </ul>
+          <li class="">
+            <app-ncte-policies></app-ncte-policies>
           </li>
           <li class="uk-margin-medium-top">
             <ul class="uk-flex-inline uk-flex-left uk-subnav uk-subnav-pill-alt uk-margin-remove"
@@ -158,13 +77,8 @@
               </li>
             </ul>
           </li>
-          <li class="uk-margin-medium-top">
-            <div class="uk-card uk-card-default uk-card-body" style="height: 700px">
-              <app-highcharts-category-map [mapData]="questionsDataArray[3]" [mapType]="'Categorization'"
-                                           [title]="'Does the country have dedicated financial contributions to the EOSC linked to the policies?'"
-                                           [pointFormat]="'{point.name}: <b>{series.name}</b>'">
-              </app-highcharts-category-map>
-            </div>
+          <li class="">
+            <app-ncte-funding></app-ncte-funding>
           </li>
           <li class="uk-margin-medium-top">
             <div class="uk-grid">
